Fix new Coverity Scan defects (run 1663736-1663755) (#1218)
Real fixes:
- grammar_rules_exprs.cc (REVERSE_NEGATIVE, CID 1663746): don't index
  predefs[f] before the f != -1 check in functional constructors.
- grammar_rules_exprs.cc (NULL_RETURNS, CID 1663740): guard the
  lookup_ident("clone_object") result before dereferencing; fall through
  to the efun form when the ident is missing.
- grammar_rules.cc (PRINTF_ARGS, CID 1663741): %d was passed a size_t;
  cast default_args_limit to int.
- lexer.l (FORWARD_NULL, CID 1663743): fail fatally if the flex buffer
  stack could not be allocated before dereferencing it in
  lpc_lex_set_source (lexer.autogen.cc regenerated).
- main_lpcshell.cc / bench_compile.cc (UNCAUGHT_EXCEPT, CIDs 1663739,
  1663742): catch std::exception at main() and report instead of
  std::terminate.

Performance (COPY_INSTEAD_OF_MOVE, CIDs 1663736/45/48/50/52):
- move instead of copy in lexer_rules_pp.cc macro-expansion guards,
  lpcshell var_names bookkeeping, and test_compiler.cc token/string
  helpers; drop the pointless stmt_body copy in lpcshell Eval entirely.

Intentional, annotated for Coverity:
- ffi.cc / main_lpcshell.cc (WRAPPER_ESCAPE, CIDs 1663754/55): the
  escaped pointers do not outlive their owners; marked coverity[].
- test_compiler.cc (SECURE_TEMP, CIDs 1663737/44/49): tmpfile() opens
  O_EXCL and unlinks immediately; kept for Windows-CI portability.

Not addressed here: the yypush_parse findings (CIDs 1663738/47/51/53)
are in bison's generated skeleton (grammar.autogen.cc), which CI
regenerates from grammar.y on every platform - they need dismissal in
the Coverity console, not a source patch.

+++ src/compiler/internal/grammar_rules.cc
@@ -219,7 +219,8 @@ void rule_func(parse_node_t** function, LPC_INT type, LPC_INT optional_star, con
           if (local.funcptr_default) {
             have_default_args = true;
             if (i > default_args_limit || argument.num_arg > default_args_limit) {
-              yyerror("Functions with default arguments can only have %d args", default_args_limit);
+              yyerror("Functions with default arguments can only have %d args",
+                      static_cast<int>(default_args_limit));
               return;
             }
             FUNCTION_DEF(fun)->min_arg--;
